requirements.txt
# PyTorch with CUDA support - インストールは別コマンドで行う
# pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu118

# Style-Bert-VITS2の依存パッケージ
aiofiles>=23.2.1
aiohappyeyeballs>=2.6.1
aiohttp>=3.11.16
annotated-types>=0.7.0
asteroid-filterbanks>=0.4.0
audioread>=3.0.1
av>=10.0.0
cmudict>=1.0.32
cn2an>=0.5.23
ctranslate2>=3.24.0
einops>=0.8.1
fastapi>=0.115.12
faster-whisper>=0.10.0
ffmpy>=0.5.0
g2p-en>=2.1.0
GPUtil>=1.4.0
gradio>=5.23.3
huggingface-hub>=0.30.1
inflect>=7.5.0
jieba>=0.42.1
julius>=0.2.7
librosa>=0.9.2
lightning>=2.5.1
loguru>=0.7.3
matplotlib>=3.10.1
nltk>=3.9.1
num2words>=0.5.14
numpy<2
omegaconf>=2.3.0
onnxruntime>=1.21.0
openai-whisper>=20240930
openai == 1.71.0
pandas>=2.2.3
pydantic>=2.11.2
pyopenjtalk-dict
pypinyin>=0.54.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
pytorch-lightning>=2.5.1
pytorch-metric-learning>=2.8.1
pyworld-prebuilt>=0.3.4.4
pyyaml>=6.0.2
safetensors>=0.5.3
scipy>=1.15.2
sentencepiece>=0.2.0
sounddevice>=0.5.1
soundfile>=0.13.1
speechbrain>=1.0.3
stable-ts>=2.19.0
tabulate>=0.9.0
tensorboard>=2.19.0
tensorboardX>=2.6.2.2
tiktoken>=0.9.0
torch-audiomentations>=0.12.0
torch_pitch_shift>=1.2.5
torchmetrics>=1.7.1
tqdm>=4.67.1
transformers>=4.39.3
typing_extensions>=4.12.2
uvicorn>=0.34.0
umap-learn==0.5.3
websockets>=15.0.0
librosa
# 開発環境用
jupyterlab>=4.3.6